src/processing/text_parser.py
import yaml
import os
from typing import Dict, List, Any

class TextParser:
    def __init__(self, config_path: str = "config/text_rules.yaml"):
        """
        Initializes the parser by loading rules from a YAML file.
        """
        if not os.path.exists(config_path):
            raise FileNotFoundError(f"Config file not found: {config_path}")
            
        with open(config_path, 'r') as f:
            self.rules = yaml.safe_load(f)
            
        self.routes = self.rules.get('routes', {})
        self.forms = self.rules.get('dosage_forms', {})

    def infer_attributes(self, text: str) -> Dict[str, str]:
        """
        Infers route and dosage form from free text descriptions.
        
        Args:
            text: Unstructured text description (e.g. "Take 1 tablet orally")
            
        Returns:
            Dict with 'route' and 'dosage_form' keys. Values are 'Unknown' if not found.
        """
        if not text or not isinstance(text, str):
            return {'route': 'Unknown', 'dosage_form': 'Unknown'}
            
        text_lower = text.lower()
        
        result = {
            'route': 'Unknown',
            'dosage_form': 'Unknown'
        }
        
        # Check Routes (Simple keyword matching)
        for route_name, keywords in self.routes.items():
            # Use word boundaries or just inclusion? 
            # Inclusion is safer for "oral tablet" but risks false positives like "coral".
            # For this challenge, simple inclusion is "reasonable".
            if any(k in text_lower for k in keywords):
                result['route'] = route_name
                break 
                
        # Check Forms
        for form_name, keywords in self.forms.items():
            if any(k in text_lower for k in keywords):
                result['dosage_form'] = form_name
                break
                
        return result

